Properly handle ASCII numerals via composed sequences
Due to the use of `unicodeScalars` previously, some ASCII characters
were being identified as emoji. In particular, the "Keycap Digit N"
characters are composed of the ASCII character followed by two other
codepoints. Keycap Digit Zero, for example, contains these scalars:

```
- [1065] : "0"
- [1066] : "\u{FE0F}"
- [1067] : "\u{20E3}"
```

In order to properly handle these sequences without false positives, we
have to split emojis into their composed character sequences and store
those as a set instead. The one complication here is that there are many
permutations of emoji with the skin tone modifiers. Instead of storing
each of these, we simply check if a character sequence has two
codepoints, and if so, that the first character is an emoji and the
second is a skin tone modifier. This is a fairly simple and efficient
way to accurately identify the presence of valid emoji.

Signed-off-by: Evan Owen <kainosnoema@gmail.com>